@@ -4,14 +4,14 @@ import 'package:flutter/material.dart';
 
 class Debouncer {
   final int milliseconds;
-  VoidCallback action;
-  Timer _timer;
+  VoidCallback? action;
+  Timer? _timer;
 
   Debouncer({this.milliseconds = 500});
 
   run(VoidCallback action) {
     if (_timer != null) {
-      _timer.cancel();
+      _timer!.cancel();
     }
     _timer = Timer(Duration(milliseconds: milliseconds), action);
   }
