@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
 class KErrorWidget extends StatelessWidget {
-  const KErrorWidget({Key key}) : super(key: key);
+  const KErrorWidget({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
