@@ -1,18 +1,14 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:listify/controller/authentication/authentication_provider.dart';
+import 'package:get/get.dart';
+import 'package:listify/controller/authentication_controller.dart';
 import 'package:listify/views/screens/home_screen.dart';
 import 'package:listify/views/screens/startup/welcome_screen.dart';
-import 'package:listify/views/widgets/k_error_widget.dart';
 
-class AuthenticationWrapper extends ConsumerWidget {
+class AuthenticationWrapper extends StatelessWidget {
   @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    final asyncUser = ref.watch(authStateChangesProvider);
-    return asyncUser.when(
-      data: (user) => user != null ? HomeScreen() : WelcomeScreen(),
-      loading: () => WelcomeScreen(),
-      error: (e, stackTrace) => KErrorWidget(),
-    );
+  Widget build(BuildContext context) {
+    return GetX<FirebaseAuthController>(builder: (authController) {
+      return authController.user.value != null ? HomeScreen() : WelcomeScreen();
+    });
   }
 }
