@@ -1,128 +1,141 @@
 import 'package:flutter/material.dart';
-import 'package:listify/controller/tasks/tasks_provider.dart';
+import 'package:get/get.dart';
+import 'package:listify/controller/task_controller.dart';
+import 'package:listify/model/todo.dart';
 import 'package:listify/services/navigation_service.dart';
 import 'package:listify/views/screens/details/details_screen.dart';
 import 'package:listify/views/styles/styles.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-class TaskCard extends ConsumerWidget {
-  final Animation<double> animation;
+class TaskCard extends StatelessWidget {
+  final Animation<double>? animation;
   final Color backgroundColor;
   final bool borderOutline;
+  final Todo todo;
 
-  TaskCard({this.animation, this.backgroundColor = KColors.white, this.borderOutline = true});
+  TaskCard(
+      {this.animation,
+      required this.todo,
+      this.backgroundColor = KColors.white,
+      this.borderOutline = true});
 
   @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    final task = ref.watch(taskProvider);
-
-    return GestureDetector(
-      onTap: () {
-        if (!task.isCompleted) {
-          ref.watch(taskDetailsProvider.state).state = task;
-          DetailsScreen().push(context);
-        }
-      },
-      child: Container(
-        margin: EdgeInsets.only(bottom: KSize.getHeight(19)),
-        child: Dismissible(
-          key: UniqueKey(),
-          background: Container(
-            padding: EdgeInsets.only(left: 20),
-            alignment: Alignment.centerLeft,
-            child: Icon(Icons.delete),
-            color: KColors.lightRed,
-          ),
-          onDismissed: (direction) async {
-            ref.read(tasksProvider).removeTodo(task.uid);
-          },
-          child: Container(
-            padding: EdgeInsets.symmetric(vertical: KSize.getHeight(15)),
-            decoration: BoxDecoration(
-              color: backgroundColor,
-              border: borderOutline ? Border.all(color: KColors.charcoal) : null,
-              borderRadius: BorderRadius.circular(4),
+  Widget build(BuildContext context) {
+    return GetBuilder<TasksController>(builder: (taskController) {
+      return GestureDetector(
+        onTap: () {
+          if (!(todo.isCompleted ?? false)) {
+            DetailsScreen(
+              todo: todo,
+            ).push(context);
+          }
+        },
+        child: Container(
+          margin: EdgeInsets.only(bottom: KSize.getHeight(19)),
+          child: Dismissible(
+            key: UniqueKey(),
+            background: Container(
+              padding: EdgeInsets.only(left: 20),
+              alignment: Alignment.centerLeft,
+              child: Icon(Icons.delete),
+              color: KColors.lightRed,
             ),
-            child: Row(
-              mainAxisAlignment: MainAxisAlignment.spaceBetween,
-              children: [
-                Flexible(
-                  child: Row(
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    children: [
-                      Flexible(
-                        child: Padding(
-                          padding: EdgeInsets.only(left: KSize.getWidth(22)),
-                          child: Column(
-                            crossAxisAlignment: CrossAxisAlignment.start,
-                            children: [
-                              Row(
-                                children: [
-                                  Padding(
-                                    padding: EdgeInsets.only(right: KSize.getWidth(22)),
-                                    child: Icon(
-                                      Icons.brightness_1_sharp,
-                                      color: task.priority == "Low"
-                                          ? Colors.green
-                                          : task.priority == "Medium"
-                                              ? Colors.orange
-                                              : Colors.red,
-                                      size: KSize.getWidth(16),
-                                    ),
-                                  ),
-                                  Flexible(
-                                    child: Text(
-                                      task.title,
-                                      style: KTextStyle.bodyText2().copyWith(fontWeight: FontWeight.w600),
-                                    ),
-                                  ),
-                                ],
-                              ),
-                              if (task.description.length > 0)
-                                Column(
+            onDismissed: (direction) async {
+              taskController.removeTodo(todo.uid);
+            },
+            child: Container(
+              padding: EdgeInsets.symmetric(vertical: KSize.getHeight(15)),
+              decoration: BoxDecoration(
+                color: backgroundColor,
+                border:
+                    borderOutline ? Border.all(color: KColors.charcoal) : null,
+                borderRadius: BorderRadius.circular(4),
+              ),
+              child: Row(
+                mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                children: [
+                  Flexible(
+                    child: Row(
+                      crossAxisAlignment: CrossAxisAlignment.start,
+                      children: [
+                        Flexible(
+                          child: Padding(
+                            padding: EdgeInsets.only(left: KSize.getWidth(22)),
+                            child: Column(
+                              crossAxisAlignment: CrossAxisAlignment.start,
+                              children: [
+                                Row(
                                   children: [
-                                    SizedBox(height: KSize.getHeight(5)),
-                                    Text(
-                                      task.description,
-                                      maxLines: 2,
-                                      overflow: TextOverflow.ellipsis,
-                                      style: KTextStyle.bodyText3(),
+                                    Padding(
+                                      padding: EdgeInsets.only(
+                                          right: KSize.getWidth(22)),
+                                      child: Icon(
+                                        Icons.brightness_1_sharp,
+                                        color: todo.priority == "Low"
+                                            ? Colors.green
+                                            : todo.priority == "Medium"
+                                                ? Colors.orange
+                                                : Colors.red,
+                                        size: KSize.getWidth(16),
+                                      ),
                                     ),
-                                    SizedBox(height: KSize.getHeight(10)),
+                                    Flexible(
+                                      child: Text(
+                                        todo.title ?? 'Unvaible',
+                                        style: KTextStyle.bodyText2().copyWith(
+                                            fontWeight: FontWeight.w600),
+                                      ),
+                                    ),
                                   ],
                                 ),
-                              Text(task.dateTime,
-                                  style: KTextStyle.bodyText2().copyWith(
-                                    color: KColors.charcoal.withOpacity(0.70),
-                                  )),
-                            ],
+                                if (todo.description != null)
+                                  if (todo.description!.length > 0)
+                                    Column(
+                                      children: [
+                                        SizedBox(height: KSize.getHeight(5)),
+                                        Text(
+                                          todo.description ?? 'Unavailable',
+                                          maxLines: 2,
+                                          overflow: TextOverflow.ellipsis,
+                                          style: KTextStyle.bodyText3(),
+                                        ),
+                                        SizedBox(height: KSize.getHeight(10)),
+                                      ],
+                                    ),
+                                Text(todo.dateTime ?? 'Unavailable',
+                                    style: KTextStyle.bodyText2().copyWith(
+                                      color: KColors.charcoal.withOpacity(0.70),
+                                    )),
+                              ],
+                            ),
                           ),
-                        ),
-                      )
-                    ],
-                  ),
-                ),
-                InkWell(
-                  onTap: () async {
-                    if (!task.isCompleted)
-                      await ref.read(tasksProvider).completeTask(task.uid);
-                    else
-                      await ref.read(tasksProvider).undoCompleteTask(task.uid);
-                  },
-                  child: Container(
-                    margin: EdgeInsets.all(KSize.getWidth(36)),
-                    child: Icon(
-                      task.isCompleted ? Icons.brightness_1 : Icons.brightness_1_outlined,
-                      color: KColors.primary,
-                      size: KSize.getWidth(24),
+                        )
+                      ],
                     ),
                   ),
-                ),
-              ],
+                  InkWell(
+                    onTap: () async {
+                      if (!(todo.isCompleted ?? false))
+                        await taskController.completeTask(todo.uid);
+                      else
+                        await taskController.undoCompleteTask(todo.uid);
+                    },
+                    child: Container(
+                      margin: EdgeInsets.all(KSize.getWidth(36)),
+                      child: Icon(
+                        todo.isCompleted ?? false
+                            ? Icons.brightness_1
+                            : Icons.brightness_1_outlined,
+                        color: KColors.primary,
+                        size: KSize.getWidth(24),
+                      ),
+                    ),
+                  ),
+                ],
+              ),
             ),
           ),
         ),
-      ),
-    );
+      );
+    });
   }
 }
